Fixed bugs where FaeUtil imported client classes on server
