+++ src/main/java/com/raggle/half_dream/mixin/EntityMixin.java
@@ -1,6 +1,7 @@
 package com.raggle.half_dream.mixin;
 
 import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
@@ -11,14 +12,18 @@ import com.raggle.half_dream.common.FaeUtil;
 
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.world.World;
 
 @Mixin(Entity.class)
 public abstract class EntityMixin {
 	
+	@Shadow
+	public abstract World getWorld();
+	
 	@Inject(method = "isInsideWall", at = @At("HEAD"), cancellable = true)
 	private void isInsideWall(CallbackInfoReturnable<Boolean> cir) {
 		if(FaeUtil.getDream((Entity)(Object)this) != 0) {
-			cir.setReturnValue(false);
+				cir.setReturnValue(false);
 		}
 	}
